01_인공지능모델운영/04_chavrusa/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
streamlit
google-generativeai
langchain
langchain-google-genai
langchain-community
faiss-cpu
pandas
